requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.2.0
python-dotenv==1.0.0
pydantic==2.5.0